rerouted some components around to pass props through multiple components, and able to display succesfully
